Update BioC XSL to include MeSH terms, chemicals
The format in which the MeSH headings and chemicals are stored in the
BioC output is described in collection.key. In short, the output
contains a third passage tag (after "title" and "abstract") called
"metadata". Each MeSH heading and Chemical is contained in one sentence
tag within the passage. This sentence, in turn, contains one (or, in
case of MeSH terms, multiple) annotation tags which, in case of
chemicals, contains its UI (unique identifier) and NameOfSubstance, and
in case of MeSH headings, one Descriptor and zero or more Qualifiers of
this descriptor. For more information on the semantics of this data,
consult the PubMed XML reference at
https://www.nlm.nih.gov/bsd/licensee/data_elements_doc.html